Disabled first two commodities in TROS form
