+++ frontend/src/features/permits/constants/constants.ts
@@ -7,18 +7,3 @@ export const VEHICLE_TYPES = [
   { value: "powerUnit", label: "Power Unit" },
   { value: "trailer", label: "Trailer" },
 ];
-
-export const TROS_PERMIT_DURATIONS = [
-  { value: 30, label: "30 Days" },
-  { value: 60, label: "60 Days" },
-  { value: 90, label: "90 Days" },
-  { value: 120, label: "120 Days" },
-  { value: 150, label: "150 Days" },
-  { value: 180, label: "180 Days" },
-  { value: 210, label: "210 Days" },
-  { value: 240, label: "240 Days" },
-  { value: 270, label: "270 Days" },
-  { value: 300, label: "300 Days" },
-  { value: 330, label: "330 Days" },
-  { value: 365, label: "1 Year" },
-];
