+++ frontend/src/features/permits/pages/TermOversize/form/ConditionsTable.tsx
@@ -11,63 +11,10 @@ import {
 } from "@mui/material";
 import { useState } from "react";
 import { Controller, useFormContext } from "react-hook-form";
-import { Commodities } from "../../../types/application";
-
-export const TROSCommodities: Commodities[] = [
-  {
-    description: "General Permit Conditions",
-    condition: "CVSE-1000",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1251",
-    checked: true,
-  },
-  {
-    description: "Permit Scope and Limitation",
-    condition: "CVSE-1070",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1261",
-    checked: true,
-  },
-  {
-    description: "Supplement for Structures",
-    condition: "CVSE-1000S",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1255",
-    checked: false,
-  },
-  {
-    description: "Log Permit Conditions",
-    condition: "CVSE-1000L",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1250",
-    checked: false,
-  },
-  {
-    description: "Routes - Woods Chips & Residual",
-    condition: "CVSE-1012",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1259",
-    checked: false,
-  },
-  {
-    description: "Restricted Routes for Hauling Wood on Wide Bunks",
-    condition: "CVSE-1013",
-    conditionLink: "https://www.th.gov.bc.ca/forms/getForm.aspx?formId=1254",
-    checked: false,
-  },
-  {
-    description: "Hay",
-    condition: "HAY (hay bales)",
-    conditionLink:
-      "https://www.cvse.ca/ctpm/com_circulars/2012/120803_comp_cc_0912_changes_hay_bales.pdf",
-    checked: false,
-  },
-  {
-    description: "Stinger Steered Transport",
-    condition: "SST (Stinger Steered Transporters)",
-    conditionLink:
-      "https://www.cvse.ca/CTPM/Com_Circulars/2017/171227-comp-cc_08-2107-Stinger-Steered-Car-Carrier.pdf",
-    checked: false,
-  },
-];
+import { TROS_COMMODITIES } from "../../../constants/termOversizeConstants";
 
 export const ConditionsTable = () => {
-  const [checkedValues, setCheckedValues] = useState(TROSCommodities);
+  const [checkedValues, setCheckedValues] = useState(TROS_COMMODITIES);
   const { control, setValue } = useFormContext();
 
   function handleSelect(checkedName: string) {
@@ -122,6 +69,7 @@ export const ConditionsTable = () => {
                           <Checkbox
                             key={row.description}
                             checked={row.checked}
+                            disabled={row.disabled}
                             onChange={() => handleSelect(row.description)}
                           />
                         );
